ResultHelpers -> TitleHelpers, refactor
